Re-organized files in Pedersen DKG
